+++ bebarball/generic/GLOBAL.py
