+++ bebarball/generic/sound.py
